autoeval_backened/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
pydantic[email]==2.9.2
email-validator==2.3.0
python-jose[cryptography]==3.5.0
bcrypt==4.2.0
python-multipart==0.0.20
pytesseract==0.3.13
transformers==4.57.3
sentence-transformers==5.1.2
torch==2.9.1
Pillow==11.0.0
PyMuPDF==1.25.0
opencv-python==4.10.0.84
numpy==2.1.3

